Reorganized Files to folow Feature Oriented Pattern
