+++ vlc/src/features/AddVirtualChannel/types/types.d.ts
@@ -1,0 +1,21 @@
+export {};
+declare global {
+  type Content = {
+    id: string;
+    title: string;
+    duration: number;
+    description: string;
+    startTime: string;
+    endTime: string;
+    image: string;
+  };
+
+  type Channel = {
+    id: string;
+    channelID: string;
+    title: string;
+    dateCreated: string;
+    author: string;
+    status: string;
+  };
+}
