+++ vlc/src/features/VirtualChannels/types/types.d.ts
@@ -1,0 +1,11 @@
+export {};
+declare global {
+  type GenericObject = { [key: string]: any };
+
+  type Table = {
+    variant: ResponsiveValue<string>;
+    caption?: String;
+    headers: String[];
+    body: GenericObject[];
+  };
+}
