+++ vlc/src/shared/themes/components/accordion.ts
@@ -1,0 +1,14 @@
+import { StyleFunctionProps } from "@chakra-ui/react";
+
+const accordion = {
+  baseStyle: (props: StyleFunctionProps) => {
+    const color: string = props.colorMode === "dark" ? "white" : "brand.800";
+    return {
+      button: {
+        color,
+      },
+    };
+  },
+};
+
+export default accordion;
