+++ vlc/src/shared/themes/foundations/breakpoints.ts
@@ -1,0 +1,10 @@
+const breakpoints = {
+  base: "0px",
+  sm: "320px",
+  md: "768px",
+  lg: "960px",
+  xl: "1200px",
+  "2xl": "1536px",
+};
+
+export default breakpoints;
