--- vlc/src/types/types.d.ts
@@ -1,22 +1,0 @@
-export {};
-
-declare global {
-  type Content = {
-    id: string;
-    title: string;
-    duration: number;
-    description: string;
-    startTime: string;
-    endTime: string;
-    image: string;
-  };
-
-  type Channel = {
-    id: string;
-    channelID: string;
-    title: string;
-    dateCreated: string;
-    author: string;
-    status: string;
-  };
-}
